requirements.txt
selenium>=4.15.0
undetected-chromedriver>=3.5.0
pandas>=2.0.0
python-dotenv>=1.0.0
colorama>=0.4.6
numpy>=1.26.0
pyautogui>=0.9.54
fake-useragent>=1.4.0
